+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.26.0 - (April 10, 2024)
+
 * Added
   * Added column span functionality to FlowsheetDataGrid.
   * Added `isAutoFocusEnabled` prop to allow auto focus of interactable elements when the only interactable element is a button or hyperlink.